Move RelatedNavigationHelper to correct directory
This commit correctly namespaces RelatedNavigationHelper and
also corrects the directory of the equivalent spec.
